+++ docs/server.py
@@ -5,6 +5,9 @@ from os import path
 import subprocess
 
 from livereload import Server
+from livereload import watcher
+
+watcher.pyinotify = None  # disable pyinotify
 
 docsdir = path.dirname(path.abspath(__file__))
 builddir = path.join(docsdir, '_build')
@@ -31,23 +34,16 @@ def docs(p: str) -> str:
 
 
 # Wtach documets
-server.watch(docs('*.rst'), cmd, delay=1)
-server.watch(docs('../*.rst'), cmd, delay=1)
 server.watch(docs('*.md'), cmd, delay=1)
-server.watch(docs('*/*.rst'), cmd, delay=1)
+server.watch(docs('../*.md'), cmd, delay=1)
+server.watch(docs('*.md'), cmd, delay=1)
 server.watch(docs('*/*.md'), cmd, delay=1)
-server.watch(docs('*/*/*.rst'), cmd, delay=1)
 server.watch(docs('*/*/*.md'), cmd, delay=1)
 
 # Watch template/style
 server.watch(docs('_templates/*.html'), cmd, delay=1)
 server.watch(docs('_static/*.css'), cmd, delay=1)
 server.watch(docs('_static/*.js'), cmd, delay=1)
-
-# Watch theme
-server.watch(docs('themes/slex/static/*.css_t'), cmd, delay=1)
-server.watch(docs('themes/slex/*.html'), cmd, delay=1)
-server.watch(docs('themes/slex/theme.conf'), cmd, delay=1)
 
 # Watch package
 server.watch(docs('../pyppeteer/*.py'), cmd, delay=1)