+++ pyppeteer/__init__.py
@@ -7,9 +7,9 @@ import logging
 
 __author__ = """Hiroyuki Takagi"""
 __email__ = 'miyako.dev@gmail.com'
-__version__ = '0.0.9'
-__chromimum_revision__ = '497674'
-__base_puppeteer_version__ = 'v0.10.2'
+__version__ = '0.0.10'
+__chromimum_revision__ = '533271'
+__base_puppeteer_version__ = 'v0.11.0'
 
 # Setup root logger
 logger = logging.getLogger('pyppeteer')
@@ -22,3 +22,7 @@ _log_handler.setLevel(logging.DEBUG)
 logger.addHandler(_log_handler)
 logger.propagate = False
 # logger.setLevel(logging.DEBUG)
+
+from pyppeteer.launcher import launch, executablePath  # noqa: E402
+
+__all__ = ['launch', 'executablePath']