+++ CHANGES.md
@@ -1,7 +1,13 @@
 History
 =======
 
-### Version 0.0.9 (next version)
+## Version 0.0.10 (2018-02-27)
+
+* Enable to import `launch` from package root
+* Change `browser.close` to coroutine function
+* Chatch up puppeteer v0.11.0
+
+### Version 0.0.9 (2017-09-09)
 
 * Delete temporary user data directory when browser closed
 * Fix bug to fail extracting zip on mac